UX: Simplified Institution Flow & Docker Production Fixes

+++ backend/lib/wealth_backend_web/controllers/user_controller.ex
@@ -7,37 +7,73 @@ defmodule WealthBackendWeb.UserController do
   action_fallback WealthBackendWeb.FallbackController
 
   def index(conn, _params) do
-    users = Accounts.list_users()
-    render(conn, :index, users: users)
+    user = conn.assigns.current_user
+    render(conn, :index, users: [user])
   end
 
   def create(conn, %{"user" => user_params}) do
+    # Default name to email (username) if not provided
+    user_params = Map.put_new(user_params, "name", user_params["email"])
+
     with {:ok, %User{} = user} <- Accounts.create_user(user_params) do
       conn
       |> put_status(:created)
-      |> put_resp_header("location", ~p"/api/users/#{user}")
+      |> WealthBackendWeb.UserAuth.log_in_user(user)
       |> render(:show, user: user)
     end
   end
 
+  def check_username(conn, %{"username" => username}) do
+    available = !Accounts.get_user_by_email(username)
+    json(conn, %{available: available})
+  end
+
   def show(conn, %{"id" => id}) do
-    user = Accounts.get_user!(id)
-    render(conn, :show, user: user)
+    user = conn.assigns.current_user
+    if to_string(user.id) == to_string(id) do
+      render(conn, :show, user: user)
+    else
+      conn
+      |> put_status(:forbidden)
+      |> put_view(json: WealthBackendWeb.ErrorJSON)
+      |> render(:"403.json")
+    end
   end
 
   def update(conn, %{"id" => id, "user" => user_params}) do
-    user = Accounts.get_user!(id)
+    user = conn.assigns.current_user
+    
+    if to_string(user.id) == to_string(id) do
+      # Automatically set tax_allowance_limit based on tax_status
+      user_params = case user_params["tax_status"] do
+        "single" -> Map.put(user_params, "tax_allowance_limit", 1000)
+        "married" -> Map.put(user_params, "tax_allowance_limit", 2000)
+        _ -> user_params
+      end
 
-    with {:ok, %User{} = user} <- Accounts.update_user(user, user_params) do
-      render(conn, :show, user: user)
+      with {:ok, %User{} = user} <- Accounts.update_user(user, user_params) do
+        render(conn, :show, user: user)
+      end
+    else
+      conn
+      |> put_status(:forbidden)
+      |> put_view(json: WealthBackendWeb.ErrorJSON)
+      |> render(:"403.json")
     end
   end
 
   def delete(conn, %{"id" => id}) do
-    user = Accounts.get_user!(id)
-
-    with {:ok, %User{}} <- Accounts.delete_user(user) do
-      send_resp(conn, :no_content, "")
+    user = conn.assigns.current_user
+    
+    if to_string(user.id) == to_string(id) do
+      with {:ok, %User{}} <- Accounts.delete_user(user) do
+        send_resp(conn, :no_content, "")
+      end
+    else
+      conn
+      |> put_status(:forbidden)
+      |> put_view(json: WealthBackendWeb.ErrorJSON)
+      |> render(:"403.json")
     end
   end
 end
